Coprocessor results should be applied to continuation.
