+++ src/coprocessor.rs
@@ -78,6 +78,8 @@ pub trait CoCircuit<F: LurkField>: Send + Sync + Clone {
 #[cfg(test)]
 pub(crate) mod test {
     use super::*;
+    use crate::circuit::gadgets::constraints::{add, mul};
+    use crate::tag::{ExprTag, Tag};
     use std::marker::PhantomData;
 
     /// A dumb Coprocessor for testing.
@@ -86,7 +88,33 @@ pub(crate) mod test {
         pub(crate) _p: PhantomData<F>,
     }
 
-    impl<F: LurkField> CoCircuit<F> for DumbCoprocessor<F> {}
+    impl<F: LurkField> CoCircuit<F> for DumbCoprocessor<F> {
+        fn arity(&self) -> usize {
+            2
+        }
+
+        fn synthesize<CS: ConstraintSystem<F>>(
+            &self,
+            cs: &mut CS,
+            _store: &Store<F>,
+            input_exprs: &[AllocatedPtr<F>],
+            input_env: &AllocatedPtr<F>,
+            input_cont: &AllocatedContPtr<F>,
+        ) -> Result<(AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>), SynthesisError>
+        {
+            let a = input_exprs[0].clone();
+            let b = &input_exprs[1];
+
+            // FIXME: Check tags.
+
+            // a^2 + b = c
+            let a2 = mul(&mut cs.namespace(|| "square"), a.hash(), a.hash())?;
+            let c = add(&mut cs.namespace(|| "add"), &a2, b.hash())?;
+            let c_ptr = AllocatedPtr::alloc_tag(cs, ExprTag::Num.to_field(), c)?;
+
+            Ok((c_ptr, input_env.clone(), input_cont.clone()))
+        }
+    }
 
     impl<F: LurkField> Coprocessor<F> for DumbCoprocessor<F> {
         /// Dumb Coprocessor takes two arguments.
